requirements.txt
chalice==1.29.0
overrides==7.4.0
pre-commit==3.4.0
pytest==7.4.2
pytest-asyncio==0.21.1
marshmallow==3.20.1
python-dotenv==1.0.0
motor==3.3.1
dependency-injector==4.41.0
boto3==1.28.77
boto3-type-annotations==0.3.1
aiohttp==3.8.6
async-lru
